Move post-chunk loaded chiseled block model updating to handler
